+++ MaterialRemove.ViewModels/Extensions/ToolHelper.cs
@@ -10,7 +10,7 @@ namespace MaterialRemove.ViewModels.Extensions
 {
     static class ToolHelper
     {
-        static internal AxisAlignedBox3d GetBound(Vector3d position, float radius, float length, Orientation orientation)
+        static internal AxisAlignedBox3d GetBound(Vector3f position, float radius, float length, Orientation orientation)
         {
             var min = position;
             var max = min;
@@ -18,28 +18,28 @@ namespace MaterialRemove.ViewModels.Extensions
             switch (orientation)
             {
                 case Orientation.XPos:
-                    min -= new Vector3d(length, radius, radius);
-                    max += new Vector3d(0.0, radius, radius);
+                    min -= new Vector3f(length, radius, radius);
+                    max += new Vector3f(0.0, radius, radius);
                     break;
                 case Orientation.XNeg:
-                    min -= new Vector3d(0.0, radius, radius);
-                    max += new Vector3d(length, radius, radius);
+                    min -= new Vector3f(0.0, radius, radius);
+                    max += new Vector3f(length, radius, radius);
                     break;
                 case Orientation.YPos:
-                    min -= new Vector3d(radius, length, radius);
-                    max += new Vector3d(radius, 0.0, radius);
+                    min -= new Vector3f(radius, length, radius);
+                    max += new Vector3f(radius, 0.0, radius);
                     break;
                 case Orientation.YNeg:
-                    min -= new Vector3d(radius, 0.0, radius);
-                    max += new Vector3d(radius, length, radius);
+                    min -= new Vector3f(radius, 0.0, radius);
+                    max += new Vector3f(radius, length, radius);
                     break;
                 case Orientation.ZPos:
-                    min -= new Vector3d(radius, radius, length);
-                    max += new Vector3d(radius, radius, 0.0);
+                    min -= new Vector3f(radius, radius, length);
+                    max += new Vector3f(radius, radius, 0.0);
                     break;
                 case Orientation.ZNeg:
-                    min -= new Vector3d(radius, radius, 0.0);
-                    max += new Vector3d(radius, radius, length);
+                    min -= new Vector3f(radius, radius, 0.0);
+                    max += new Vector3f(radius, radius, length);
                     break;
                 default:
                     throw new ArgumentOutOfRangeException();
